+dao +svc member

+++ src/main/java/com/kh/great/domain/Member.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 @Data
 @NoArgsConstructor
 public class Member {
-  private Long memNumber;                   //varchar2(9)
+  private int memNumber;                   //varchar2(9)
   private String memType;                     //varchar2(15)
   private String memId;                       //varchar2(30)
   private String memPassword;                 //varchar2(18)
